requirements.txt
argparse
pyinstaller
openai-whisper==20231117
